diffCtrlApp/Db/diff_ctrl.db: Add speed and acceleration protection
There is a possibility that the motors hit each other if the motors move at different speeds or with different acceleration.
In order to avoid this problem, the speed of both motors are set to the same value. The left edge's speed and acceleration are used as reference. The right edge's speed and acceleration are kept equal to left's.

The following parameters of the motor records in the low level IOCs are controlled:

        -VELO
        -VMAX
        -ACCL