Add inner limits
Add PVs "NegEdgeInnerLim-SP/RB" and "PosEdgeInnerLim-SP/RB" to define inner limits to the Slit Motion. These limits can be, for instance, the position of hard stops that limit the closing movement of the slit. The inner limits allow for negative width, differently from the previous IOC version configuration, where the minimum width was fixed in zero.