Add conversion functions to *App/src/
Add custom conversion functions for Slit and Scraper
applications in the *App/src/ directory.

The functions used will depend on the configurations
specified in the IOC startup (not implemented yet).

There are separate conversion functions for setpoint
and readback. This helps to isolate problems in
conversion functions and prevents failure in one
'direction' of conversion to impact conversion in
the other.

Each blade or actuator of a Slit or Scraper can have
different conversion functions. Currently, only the
Scraper needs different functions for each blade.