Add new IOC argument: device type
Add a new IOC argument, DEVICE, which specifies which of the available
conversion functions and parameters are goig to be used.

Currently, there are 3 options: slit, scraper-h, and scraper-v.

The runDiffCtrl.sh script sets the configuration macros according to
the selected device type.

OBS: The database file does not support different conversions yet.