.../myUtil.h: Modify VScraper Top blade conversion
Flip the sign of the Y array of the VScraper Top blade conversion table.

The points used in this table were acquired in a reference frame common
to the lower blade. The ref. frames of the data provided for
coordinate conversion should be opposite though. Each frame should point
from the corresponding blade towards the Scraper center plane.

To fix this, the sign of the array elements were flipped.