Update Edge Position SP after homing
Force set point positions equal to readback values for
both negative and positive edge, without moving the motors.

This is important in order to update the fields of the transform record
'CoordinateConv'. Since this record updates the 'center', 'width',
'positive edge position' and 'negative edge position' every time
a new set point is set, NOT updating these fields after a homing
could cause an edge to move to an incorrect position after one
of the other parameters is set to a new position. This happens
because the fields of the transform record still hold the old
values, and thus, "does not know that the edges have moved".

For instance, the new edges position after homing could result
in a different center than before. If the user tries to change
the width, the old center value would be considered, which is
probably not what the user wanted.

Updating the set points also updates the transform record fields,
solving the problem.