Change PV properties names according to device type
Change the names of the PVs related to a specific edge
according to the device type.

For instance, for the Vertical Scraper, the edges positions
are given by:
 - BottomEdgePos-RB
 - TopEdgePos-RB

On the other hand, for the Horizontal Scraper, they are given by:

 - InnerEdgePos-RB
 - OuterEdgePos-RB

Other PVs that are directly associated to an edge also
have been updated to contain an intuitive reference
to the respective edge.